+++ siba/03-callback-args_test.go
@@ -1,18 +1,18 @@
 // Copyright (C) 2025 azumakuniyuki and sisimai development team, All rights reserved.
 // This software is distributed under The BSD 2-Clause License.
-package sis
+package siba
 
-//  _____         _      __   _       ____      _ _ _                _        _                  
-// |_   _|__  ___| |_   / /__(_)___  / ___|__ _| | | |__   __ _  ___| | __   / \   _ __ __ _ ___ 
-//   | |/ _ \/ __| __| / / __| / __|| |   / _` | | | '_ \ / _` |/ __| |/ /  / _ \ | '__/ _` / __|
-//   | |  __/\__ \ |_ / /\__ \ \__ \| |__| (_| | | | |_) | (_| | (__|   <  / ___ \| | | (_| \__ \
-//   |_|\___||___/\__/_/ |___/_|___(_)____\__,_|_|_|_.__/ \__,_|\___|_|\_\/_/   \_\_|  \__, |___/
-//                                                                                     |___/     
+//  _____         _      __   _ _             ____      _ _ _                _        _                  
+// |_   _|__  ___| |_   / /__(_) |__   __ _  / ___|__ _| | | |__   __ _  ___| | __   / \   _ __ __ _ ___ 
+//   | |/ _ \/ __| __| / / __| | '_ \ / _` || |   / _` | | | '_ \ / _` |/ __| |/ /  / _ \ | '__/ _` / __|
+//   | |  __/\__ \ |_ / /\__ \ | |_) | (_| || |__| (_| | | | |_) | (_| | (__|   <  / ___ \| | | (_| \__ \
+//   |_|\___||___/\__/_/ |___/_|_.__/ \__,_(_)____\__,_|_|_|_.__/ \__,_|\___|_|\_\/_/   \_\_|  \__, |___/
+//                                                                                             |___/     
 import "testing"
 
 func TestCallbackArg0(t *testing.T) {
 	cc := "CallbackArg0"
-	fn := "sis.CallbackArg0"
+	fn := "siba.CallbackArg0"
 	cv := &CallbackArg0{
 		Headers: map[string][]string{"Nekochan": []string{"Kijitora", "Michistuna"}},
 		Payload: &cc,
@@ -27,7 +27,7 @@ func TestCallbackArg0(t *testing.T) {
 
 func TestCallbackArg1(t *testing.T) {
 	cc := "CallbackArg1"
-	fn := "sis.CallbackArg1"
+	fn := "siba.CallbackArg1"
 	cv := &CallbackArg1{
 		Path: "/var/mail/root",
 		Kind: "mailbox",
