+++ siba/05-delivery-matter_test.go
@@ -1,13 +1,13 @@
 // Copyright (C) 2025 azumakuniyuki and sisimai development team, All rights reserved.
 // This software is distributed under The BSD 2-Clause License.
-package sis
+package siba
 
-//  _____         _      __   _       ____       _ _                      __  __       _   _            
-// |_   _|__  ___| |_   / /__(_)___  |  _ \  ___| (_)_   _____ _ __ _   _|  \/  | __ _| |_| |_ ___ _ __ 
-//   | |/ _ \/ __| __| / / __| / __| | | | |/ _ \ | \ \ / / _ \ '__| | | | |\/| |/ _` | __| __/ _ \ '__|
-//   | |  __/\__ \ |_ / /\__ \ \__ \_| |_| |  __/ | |\ V /  __/ |  | |_| | |  | | (_| | |_| ||  __/ |   
-//   |_|\___||___/\__/_/ |___/_|___(_)____/ \___|_|_| \_/ \___|_|   \__, |_|  |_|\__,_|\__|\__\___|_|   
-//                                                                  |___/                               
+//  _____         _      __   _ _             ____       _ _                      __  __       _   _            
+// |_   _|__  ___| |_   / /__(_) |__   __ _  |  _ \  ___| (_)_   _____ _ __ _   _|  \/  | __ _| |_| |_ ___ _ __ 
+//   | |/ _ \/ __| __| / / __| | '_ \ / _` | | | | |/ _ \ | \ \ / / _ \ '__| | | | |\/| |/ _` | __| __/ _ \ '__|
+//   | |  __/\__ \ |_ / /\__ \ | |_) | (_| |_| |_| |  __/ | |\ V /  __/ |  | |_| | |  | | (_| | |_| ||  __/ |   
+//   |_|\___||___/\__/_/ |___/_|_.__/ \__,_(_)____/ \___|_|_| \_/ \___|_|   \__, |_|  |_|\__,_|\__|\__\___|_|   
+//                                                                          |___/                               
 import "testing"
 import "strings"
