+++ siba/06-not-decoded_test.go
@@ -1,12 +1,12 @@
 // Copyright (C) 2024-2025 azumakuniyuki and sisimai development team, All rights reserved.
 // This software is distributed under The BSD 2-Clause License.
-package sis
+package siba
 
-//  _____         _      __   _       _   _       _   ____                     _          _ 
-// |_   _|__  ___| |_   / /__(_)___  | \ | | ___ | |_|  _ \  ___  ___ ___   __| | ___  __| |
-//   | |/ _ \/ __| __| / / __| / __| |  \| |/ _ \| __| | | |/ _ \/ __/ _ \ / _` |/ _ \/ _` |
-//   | |  __/\__ \ |_ / /\__ \ \__ \_| |\  | (_) | |_| |_| |  __/ (_| (_) | (_| |  __/ (_| |
-//   |_|\___||___/\__/_/ |___/_|___(_)_| \_|\___/ \__|____/ \___|\___\___/ \__,_|\___|\__,_|
+//  _____         _      __   _ _             _   _       _   ____                     _          _ 
+// |_   _|__  ___| |_   / /__(_) |__   __ _  | \ | | ___ | |_|  _ \  ___  ___ ___   __| | ___  __| |
+//   | |/ _ \/ __| __| / / __| | '_ \ / _` | |  \| |/ _ \| __| | | |/ _ \/ __/ _ \ / _` |/ _ \/ _` |
+//   | |  __/\__ \ |_ / /\__ \ | |_) | (_| |_| |\  | (_) | |_| |_| |  __/ (_| (_) | (_| |  __/ (_| |
+//   |_|\___||___/\__/_/ |___/_|_.__/ \__,_(_)_| \_|\___/ \__|____/ \___|\___\___/ \__,_|\___|\__,_|
 import "testing"
 import "strings"
