+++ siba/07-rising-underway_test.go
@@ -1,13 +1,13 @@
 // Copyright (C) 2025 azumakuniyuki and sisimai development team, All rights reserved.
 // This software is distributed under The BSD 2-Clause License.
-package sis
+package siba
 
-//  _____         _      __   _       ____  _     _             _   _           _                               
-// |_   _|__  ___| |_   / /__(_)___  |  _ \(_)___(_)_ __   __ _| | | |_ __   __| | ___ _ ____      ____ _ _   _ 
-//   | |/ _ \/ __| __| / / __| / __| | |_) | / __| | '_ \ / _` | | | | '_ \ / _` |/ _ \ '__\ \ /\ / / _` | | | |
-//   | |  __/\__ \ |_ / /\__ \ \__ \_|  _ <| \__ \ | | | | (_| | |_| | | | | (_| |  __/ |   \ V  V / (_| | |_| |
-//   |_|\___||___/\__/_/ |___/_|___(_)_| \_\_|___/_|_| |_|\__, |\___/|_| |_|\__,_|\___|_|    \_/\_/ \__,_|\__, |
-//                                                        |___/                                           |___/ 
+//  _____         _      __   _ _             ____  _     _             _   _           _                               
+// |_   _|__  ___| |_   / /__(_) |__   __ _  |  _ \(_)___(_)_ __   __ _| | | |_ __   __| | ___ _ ____      ____ _ _   _ 
+//   | |/ _ \/ __| __| / / __| | '_ \ / _` | | |_) | / __| | '_ \ / _` | | | | '_ \ / _` |/ _ \ '__\ \ /\ / / _` | | | |
+//   | |  __/\__ \ |_ / /\__ \ | |_) | (_| |_|  _ <| \__ \ | | | | (_| | |_| | | | | (_| |  __/ |   \ V  V / (_| | |_| |
+//   |_|\___||___/\__/_/ |___/_|_.__/ \__,_(_)_| \_\_|___/_|_| |_|\__, |\___/|_| |_|\__,_|\___|_|    \_/\_/ \__,_|\__, |
+//                                                                |___/                                           |___/ 
 import "testing"
 
 // Digest []DeliveryMatter // List of DeliveryMatter structs
