+++ siba/08-email-address_test.go
@@ -1,12 +1,12 @@
 // Copyright (C) 2025 azumakuniyuki and sisimai development team, All rights reserved.
 // This software is distributed under The BSD 2-Clause License.
-package sis
+package siba
 
-//  _____         _      __   _       _____                 _ _    _       _     _                   
-// |_   _|__  ___| |_   / /__(_)___  | ____|_ __ ___   __ _(_) |  / \   __| | __| |_ __ ___  ___ ___ 
-//   | |/ _ \/ __| __| / / __| / __| |  _| | '_ ` _ \ / _` | | | / _ \ / _` |/ _` | '__/ _ \/ __/ __|
-//   | |  __/\__ \ |_ / /\__ \ \__ \_| |___| | | | | | (_| | | |/ ___ \ (_| | (_| | | |  __/\__ \__ \
-//   |_|\___||___/\__/_/ |___/_|___(_)_____|_| |_| |_|\__,_|_|_/_/   \_\__,_|\__,_|_|  \___||___/___/
+//  _____         _      __   _ _             _____                 _ _    _       _     _                   
+// |_   _|__  ___| |_   / /__(_) |__   __ _  | ____|_ __ ___   __ _(_) |  / \   __| | __| |_ __ ___  ___ ___ 
+//   | |/ _ \/ __| __| / / __| | '_ \ / _` | |  _| | '_ ` _ \ / _` | | | / _ \ / _` |/ _` | '__/ _ \/ __/ __|
+//   | |  __/\__ \ |_ / /\__ \ | |_) | (_| |_| |___| | | | | | (_| | | |/ ___ \ (_| | (_| | | |  __/\__ \__ \
+//   |_|\___||___/\__/_/ |___/_|_.__/ \__,_(_)_____|_| |_| |_|\__,_|_|_/_/   \_\__,_|\__,_|_|  \___||___/___/
 import "testing"
 import "strings"
