+++ siba/09-fact_test.go
@@ -1,12 +1,12 @@
 // Copyright (C) 2025 azumakuniyuki and sisimai development team, All rights reserved.
 // This software is distributed under The BSD 2-Clause License.
-package sis
+package siba
 
-//  _____         _      __   _       _____          _   
-// |_   _|__  ___| |_   / /__(_)___  |  ___|_ _  ___| |_ 
-//   | |/ _ \/ __| __| / / __| / __| | |_ / _` |/ __| __|
-//   | |  __/\__ \ |_ / /\__ \ \__ \_|  _| (_| | (__| |_ 
-//   |_|\___||___/\__/_/ |___/_|___(_)_|  \__,_|\___|\__|
+//  _____         _      __   _ _             _____          _   
+// |_   _|__  ___| |_   / /__(_) |__   __ _  |  ___|_ _  ___| |_ 
+//   | |/ _ \/ __| __| / / __| | '_ \ / _` | | |_ / _` |/ __| __|
+//   | |  __/\__ \ |_ / /\__ \ | |_) | (_| |_|  _| (_| | (__| |_ 
+//   |_|\___||___/\__/_/ |___/_|_.__/ \__,_(_)_|  \__,_|\___|\__|
 import "testing"
 
 func TestFact(t *testing.T) {
